chore: 不要ファイルの削除と整理
