+++ .github/workflows/scrape_prices.yml
@@ -10,7 +10,6 @@ env:
   AWS_REGION: ap-northeast-1
   PYTHONPATH: ${{ github.workspace }}
   PYTHONUNBUFFERED: 1  # Pythonの出力バッファリングを無効化
-  TEST_ENV: true  # テスト環境フラグ
 
 jobs:
   scrape:
@@ -61,49 +60,3 @@ jobs:
         if: steps.scrape.outcome == 'failure'
         run: |
           python scripts/price_scraping/send_error_notification.py "${{ steps.scrape.outcome }}" "${{ steps.scrape.outputs.error_message }}"
-
-      - name: Run tests
-        if: ${{ github.event_name == 'workflow_dispatch' }}
-        env:
-          AWS_REGION: ap-northeast-1
-          TEST_ENV: true
-          PYTHONPATH: ${{ github.workspace }}
-          PYTHONUNBUFFERED: 1
-          CONFIG_FILE: config/config.testing.yaml
-        run: |
-          # テスト用設定ファイルの作成
-          mkdir -p config
-          cat > config/config.testing.yaml << 'EOL'
-          app:
-            debug: true
-            log_level: DEBUG
-            secret_key: test-secret-key-16ch
-
-          scraper:
-            kaitori_rudea_urls:
-              - "https://kaitori-rudeya.com/category/detail/183"  # iPhone 16
-              - "https://kaitori-rudeya.com/category/detail/185"  # iPhone 16 Pro
-              - "https://kaitori-rudeya.com/category/detail/186"  # iPhone 16 Pro Max
-              - "https://kaitori-rudeya.com/category/detail/205"  # iPhone 16 e
-            apple_store_url: "https://www.apple.com/jp/shop/buy-iphone"
-            request_timeout: 30
-            retry_count: 3
-            user_agent: "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/91.0.4472.124 Safari/537.36"
-          EOL
-          
-          # テストの実行
-          pip install pytest pytest-cov pytest-asyncio
-          python -m pytest tests/ --cov=src --cov-report=xml -v --tb=short --capture=no 2>&1 | tee test_output.log
-          
-          # テスト結果の確認
-          if [ $? -ne 0 ]; then
-            echo "Tests failed. Check test_output.log for details."
-            exit 1
-          fi
-
-      - name: Upload coverage report
-        if: ${{ github.event_name == 'workflow_dispatch' }}
-        uses: codecov/codecov-action@v4
-        with:
-          file: ./coverage.xml
-          fail_ci_if_error: true
