--- config/config.testing.yaml
@@ -1,43 +1,0 @@
-app:
-  debug: true
-  log_level: DEBUG
-  secret_key: test-secret-key-16ch
-
-scraper:
-  selectors:
-    kaitori:
-      item: ".price-item"
-      model: ".model"
-      price: ".price"
-      condition: ".condition"
-    official:
-      item: ".as-producttile"
-      model: ".as-producttile-title"
-      price: ".as-price-currentprice"
-      condition: ".as-producttile-condition"
-  headers:
-    User-Agent: "Mozilla/5.0 (iPhone; CPU iPhone OS 14_7_1 like Mac OS X) AppleWebKit/605.1.15 (KHTML, like Gecko) Version/14.1.2 Mobile/15E148 Safari/604.1"
-  max_retries: 3
-  timeout: 30
-  cache_duration: 3600 # 1 hour in seconds
-  cache_dir: "./test_cache"
-
-urls:
-  kaitori:
-    - "https://example.com/kaitori/iphone"
-  official:
-    - "https://example.com/official/iphone"
-
-line:
-  channel_access_token: "test_access_token"
-  channel_secret: "test_secret"
-
-dynamodb:
-  table_name: "test_prices_table"
-  region: "ap-northeast-1"
-  endpoint_url: "http://localhost:8000"
-
-logging:
-  level: "DEBUG"
-  format: "%(asctime)s - %(name)s - %(levelname)s - %(message)s"
-  file: "./logs/test.log"
