--- lambdas/get_prices_lambda/requirements.txt
@@ -1,4 +1,0 @@
-requests==2.31.0
-beautifulsoup4==4.12.2
-PyYAML==6.0.1
-boto3==1.38.13
