+++ requirements-base.txt
@@ -1,4 +1,8 @@
-urllib3>=2.0.5,<3.0
-requests>=2.32.3,<3.0
 boto3>=1.34.0
-python-dotenv==1.1.0 
+requests>=2.32.3
+python-dotenv>=1.1.0
+beautifulsoup4>=4.13.4
+pyyaml>=6.0.0
+lxml>=5.1.0
+aiohttp>=3.9.0
+tenacity>=8.2.0
