--- requirements-test.txt
@@ -1,3 +1,0 @@
-pytest==7.4.4
-pytest-cov==6.1.1
-pytest-mock==3.14.0
