+++ requirements-web.txt
@@ -1,5 +1,10 @@
-selenium>=4.16.0,<5.0
-playwright==1.52.0
-beautifulsoup4==4.13.4
-flask==3.1.0
-gunicorn==23.0.0 
+- selenium>=4.16.0,<5.0
+- playwright==1.52.0
+- beautifulsoup4==4.13.4
+- flask==3.1.0
+- gunicorn==23.0.0
++ -r requirements-base.txt
++ selenium>=4.16.0,<5.0
++ flask==3.1.0
++ gunicorn==23.0.0
++ playwright==1.52.0
