+++ requirements.txt
@@ -1,13 +1,16 @@
-boto3>=1.34.0
-requests==2.31.0
-line-bot-sdk>=3.9.0
-playwright>=1.42.0
-pytest>=8.2.0
-pytest-asyncio>=0.23.0
-pytest-cov>=4.1.0
-python-dotenv>=1.1.0
-pyyaml>=6.0.0
-aiohttp>=3.9.0
-tenacity>=8.2.0
-beautifulsoup4>=4.13.4
-lxml>=5.1.0
+- boto3>=1.34.0
+- requests==2.31.0
+- line-bot-sdk>=3.9.0
+- playwright>=1.42.0
+- pytest>=8.2.0
+- pytest-asyncio>=0.23.0
+- pytest-cov>=4.1.0
+- python-dotenv>=1.1.0
+- pyyaml>=6.0.0
+- aiohttp>=3.9.0
+- tenacity>=8.2.0
+- beautifulsoup4>=4.13.4
+- lxml>=5.1.0
++ -r requirements-base.txt
++ line-bot-sdk>=3.9.0
++ playwright>=1.42.0
