--- src/lambda_functions/get_prices_lambda/config/__init__.py
@@ -1,3 +1,0 @@
-"""
-Configuration management for the price comparison application.
-"""
