--- src/lambda_functions/get_prices_lambda/config/config.yaml
@@ -1,26 +1,0 @@
-scraper:
-  selectors:
-    price_item: ".price-item"
-    model: ".model"
-    price: ".price"
-    condition: ".condition"
-  headers:
-    User-Agent: "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/91.0.4472.114 Safari/537.36"
-    Accept: "text/html,application/xhtml+xml,application/xml;q=0.9,image/webp,*/*;q=0.8"
-    Accept-Language: "ja,en-US;q=0.7,en;q=0.3"
-    Connection: "keep-alive"
-    Upgrade-Insecure-Requests: "1"
-  user_agent: "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/91.0.4472.114 Safari/537.36"
-  max_retries: 3
-  timeout: 30
-  cache_duration: 3600
-
-urls:
-  kaitori:
-    - "https://example.com/kaitori"
-  official:
-    - "https://example.com/official"
-
-dynamodb:
-  table_name: "price_data"
-  region: "ap-northeast-1"
