--- terraform/requirements.txt
@@ -1,10 +1,0 @@
-boto3==1.38.13
-requests==2.32.3
-line-bot-sdk==3.17.1
-beautifulsoup4==4.13.4
-flask==3.1.0
-aiohttp==3.11.18
-attrs==25.3.0
-async-timeout==5.0.1
-aiosignal==1.3.2
-aiohappyeyeballs==2.6.1
