+++ terraform/scripts/create_patch.py
