--- terraform/scripts/requirements.txt
@@ -1,6 +1,0 @@
-boto3>=1.26.0
-requests>=2.31.0
-beautifulsoup4==4.12.0
-PyYAML==6.0.1
-python-dateutil==2.9.0.post0
-pytz==2024.2
