--- tests/__init__.py
@@ -1,1 +1,0 @@
-# This file makes the tests directory a Python package
