--- tests/pytest.ini
@@ -1,13 +1,0 @@
-[pytest]
-pythonpath = .
-testpaths = tests/unit
-python_files = test_*.py
-addopts = -v -s --cov=src --cov-report=term-missing --cov-report=xml --tb=short
-markers =
-    unit: Unit tests
-    integration: Integration tests
-filterwarnings =
-    ignore::DeprecationWarning
-    ignore::UserWarning
-log_cli = true
-log_cli_level = INFO
