--- tests/unit/__init__.py
@@ -1,1 +1,0 @@
-# This file makes the unit directory a Python package
